feat(world-protocol): 3.4.3 opcode enum + split table, handlers stubbed (brick B1)
Replace AzerothCore's legacy `enum Opcodes : uint16` + flat opcode table with
TrinityCore wotlk_classic's split sparse `OpcodeClient`/`OpcodeServer : uint32`
enums (TC 61581 / 3.4.3 values) and a dual `_internalTableClient` /
`_internalTableServer` handler table indexed via `GetOpcodeArrayIndex`.

Opcodes.h:
- Keep brick-A `enum ConnectionType` + `UNKNOWN_OPCODE`.
- Add `OpcodeClient`/`OpcodeServer` enum bodies verbatim from the TC reference
  (incl. `= UNKNOWN_OPCODE` aliases), both `GetOpcodeArrayIndex` overloads,
  `NUM_CMSG_OPCODES = 1420`, `NUM_SMSG_OPCODES = 1454`, and `IsInstanceOnlyOpcode`.
- Keep AC's virtual `OpcodeHandler`/`ClientOpcodeHandler` + `PacketHandler`
  dispatch design; add `ConnectionType ConnectionIndex` to `ServerOpcodeHandler`.
- New `OpcodeTable` with split tables, `IsValid()`/`operator[]` overloads, and
  `GetOpcodeNameForLogging(OpcodeClient|OpcodeServer)` (6-hex-digit format).

Opcodes.cpp:
- Keep AC's `PacketHandler`/`get_packet_class` machinery + `DEFINE_HANDLER` /
  `DEFINE_SERVER_OPCODE_HANDLER` macros; adapt `ValidateAndSet*` to the split
  tables + `GetOpcodeArrayIndex` + `UNKNOWN_OPCODE` checks.
- Register the handshake/essential CMSGs to their real early-process path
  (CMSG_AUTH_SESSION, CMSG_AUTH_CONTINUED_SESSION, CMSG_ENTER_ENCRYPTED_MODE_ACK,
  CMSG_KEEP_ALIVE, CMSG_LOG_DISCONNECT, CMSG_PING -> Handle_EarlyProccess);
  every other CMSG (668) is a STATUS_UNHANDLED/Handle_NULL stub. Existing AC
  Handle* methods are intentionally NOT rebound (deferred to B2+). SMSG lines
  (1125) ported essentially verbatim.

Direct table consumers made null-safe (unregistered opcodes are now legitimately
nullptr): WorldSession::Update() dispatch, MapSessionFilter/WorldSessionFilter,
the AntiDos flood log, and WorldSocket::ReadDataHandler all guard with
`opcodeTable.IsValid()` + null checks. Brick-A landmines widened to uint32:
`GetAntiDosPolicyForOpcode` + `_PacketThrottlingMap` + the `antidos_opcode_policies`
loader (now indexed via `GetOpcodeArrayIndex`), and
`ClientPktHeader::IsValidOpcode()`. A few renamed handshake-path opcode refs in
the consumer files updated to their 3.4.3 names (CMSG_TIME_SYNC_RESPONSE,
CMSG_WARDEN3_DATA, CMSG_ENUM_CHARACTERS).

This is an intentional foundation checkpoint: the `game` target does NOT fully
build yet. Replacing the enum breaks ~673 opcode-name references plus ~75 uses of
the removed `Opcodes` type across ~100 other files (handlers, packet structs,
Unit/Player, scripts); those are fixed per-feature in brick B2+. The opcode
system itself (Opcodes.h/.cpp) and the direct table consumers are complete and
internally consistent (validated: all concrete opcodes map to unique in-range
array slots).

+++ src/server/game/Server/WorldSession.h
@@ -1206,7 +1206,8 @@ protected:
     protected:
         WorldSession* Session;
     private:
-        typedef std::unordered_map<uint16, PacketCounter> PacketThrottlingMap;
+        // brick B: opcodes are uint32 (OpcodeClient) for the 3.4.3 wire protocol
+        typedef std::unordered_map<uint32, PacketCounter> PacketThrottlingMap;
         // mark this member as "mutable" so it can be modified even in const functions
         mutable PacketThrottlingMap _PacketThrottlingMap;
